fix: change type to uint160

+++ src/SablierV2ProxyTarget.sol
@@ -166,7 +166,7 @@ contract SablierV2ProxyTarget is
         // Calculate the sum of all of stream amounts. It is safe to use unchecked addition because one of the create
         // transactions will revert if there is overflow.
         uint256 i;
-        uint128 transferAmount;
+        uint160 transferAmount;
         for (i = 0; i < batchSize;) {
             unchecked {
                 transferAmount += batch[i].totalAmount;
@@ -221,7 +221,7 @@ contract SablierV2ProxyTarget is
         // Calculate the sum of all of stream amounts. It is safe to use unchecked addition because one of the create
         // transactions will revert if there is overflow.
         uint256 i;
-        uint128 transferAmount;
+        uint160 transferAmount;
         for (i = 0; i < batchSize;) {
             unchecked {
                 transferAmount += batch[i].totalAmount;
@@ -392,7 +392,7 @@ contract SablierV2ProxyTarget is
         // Calculate the sum of all of stream amounts. It is safe to use unchecked addition because one of the create
         // transactions will revert if there is overflow.
         uint256 i;
-        uint128 transferAmount;
+        uint160 transferAmount;
         for (i = 0; i < batchSize;) {
             unchecked {
                 transferAmount += batch[i].totalAmount;
@@ -447,7 +447,7 @@ contract SablierV2ProxyTarget is
         // Calculate the sum of all of stream amounts. It is safe to use unchecked addition because one of the create
         // transactions will revert if there is overflow.
         uint256 i;
-        uint128 transferAmount;
+        uint160 transferAmount;
         for (i = 0; i < batchSize;) {
             unchecked {
                 transferAmount += batch[i].totalAmount;
